main-2

+++ MyBankServer/MyBank.Persistence/Repositories/AdminRepository.cs
@@ -4,6 +4,7 @@ public class AdminRepository : IAdminRepository
 {
     private readonly MyBankDbContext _dbContext;
     private readonly IMapper _mapper;
+
     public AdminRepository(MyBankDbContext dbContext, IMapper mapper)
     {
         _dbContext = dbContext;
@@ -21,8 +22,8 @@ public class AdminRepository : IAdminRepository
 
     public async Task<Admin> GetById(int id)
     {
-        var adminEntity = await _dbContext.Admins
-            .AsNoTracking()
+        var adminEntity = await _dbContext
+            .Admins.AsNoTracking()
             .FirstOrDefaultAsync(а => а.Id == id);
 
         return _mapper.Map<Admin>(adminEntity);
@@ -30,8 +31,8 @@ public class AdminRepository : IAdminRepository
 
     public async Task<Admin> GetByLogin(string login)
     {
-        var adminEntity = await _dbContext.Admins
-            .AsNoTracking()
+        var adminEntity = await _dbContext
+            .Admins.AsNoTracking()
             .FirstOrDefaultAsync(а => а.Login == login);
 
         return _mapper.Map<Admin>(adminEntity);
@@ -39,28 +40,23 @@ public class AdminRepository : IAdminRepository
 
     public async Task<List<Admin>> GetAll()
     {
-        var adminEntitiesList = await _dbContext.Admins
-            .AsNoTracking()
-            .ToListAsync();
+        var adminEntitiesList = await _dbContext.Admins.AsNoTracking().ToListAsync();
 
         return _mapper.Map<List<Admin>>(adminEntitiesList);
     }
 
     public async Task<bool> UpdateInfo(int id, string nickname)
     {
-        var number = await _dbContext.Admins
-            .Where(a => a.Id == id)
-            .ExecuteUpdateAsync(s => s
-                .SetProperty(a => a.Nickname, nickname));
+        var number = await _dbContext
+            .Admins.Where(a => a.Id == id)
+            .ExecuteUpdateAsync(s => s.SetProperty(a => a.Nickname, nickname));
 
         return (number == 0) ? false : true;
     }
 
     public async Task<bool> Delete(int id)
     {
-        var number = await _dbContext.Admins
-            .Where(a => a.Id == id)
-            .ExecuteDeleteAsync();
+        var number = await _dbContext.Admins.Where(a => a.Id == id).ExecuteDeleteAsync();
 
         return (number == 0) ? false : true;
     }
